% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active.R
\name{ActiveConditions}
\alias{ActiveConditions}
\title{Return list of active conditions}
\usage{
ActiveConditions()
}
\value{
The active conditions, checks for errors
}
\description{
Return list of active conditions
}
