% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active.R
\name{ActiveExperiments}
\alias{ActiveExperiments}
\title{Returns a list of the active experiments}
\usage{
ActiveExperiments()
}
\value{
List of active experiments
}
\description{
Returns a list of the active experiments
}
