% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{ArchiveMPC}
\alias{ArchiveMPC}
\title{Archive the .999 files to the data/mpc directory}
\usage{
ArchiveMPC(mpcpath)
}
\arguments{
\item{mpcpath}{The path to the data/mpc directory (write to json first!)}
}
\description{
Archive the .999 files to the data/mpc directory
}
