% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{CreateActiveExperiment}
\alias{CreateActiveExperiment}
\title{Creates an active experiment given the critical files}
\usage{
CreateActiveExperiment(name, dropbox, conditionsFile, protocolFile,
  eventcodesFile = NULL)
}
\arguments{
\item{name}{The name of the experiment (e.g., "Consistent Weber)}

\item{dropbox}{The dropbox path}

\item{conditionsFile}{The conditions file path}

\item{protocolFile}{The protocol file path}

\item{eventcodesFile}{The eventcodes file path}
}
\value{
NULL (creates the experiment directory structure)
}
\description{
Creates an active experiment given the critical files
}
