% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{CreateAnalysisMainFile}
\alias{CreateAnalysisMainFile}
\title{Creates an active experiment given the critical files}
\usage{
CreateAnalysisMainFile(file)
}
\arguments{
\item{file}{The filename of the new analysis file (e.g., ~/Desktop/h_main.R)}
}
\value{
NULL (creates the analysis file)
}
\description{
Creates an active experiment given the critical files
}
