% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{CreateManuscriptMainFile}
\alias{CreateManuscriptMainFile}
\title{Creates the manuscript's main Markdown file}
\usage{
CreateManuscriptMainFile(file)
}
\arguments{
\item{file}{The filename of the new manuscript file (e.g., ~/Dropbox/Active/h_consisten_weber/manuscript/h_main.md)}
}
\value{
NULL (creates the markdown file)
}
\description{
Creates the manuscript's main Markdown file
}
