% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxtest.R
\name{CumulativeBlockedDeliveries}
\alias{CumulativeBlockedDeliveries}
\title{Obtain and plot the cumulative number of deliveries while the feeder is blocked}
\usage{
CumulativeBlockedDeliveries(df, return_ax = FALSE)
}
\arguments{
\item{df}{timestamped data file}
}
\value{
df same timestamped file
}
\description{
Obtain and plot the cumulative number of deliveries while the feeder is blocked
}
