% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxtest.R
\name{CumulativeFoodAmount}
\alias{CumulativeFoodAmount}
\title{Obtain and plot the cumulative food amounts}
\usage{
CumulativeFoodAmount(df, return_ax = FALSE)
}
\arguments{
\item{df}{timestamped data file}
}
\value{
df same timestamped file
}
\description{
Obtain and plot the cumulative food amounts
}
