% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{DataListForJson}
\alias{DataListForJson}
\title{Return a list of the data ready for json}
\usage{
DataListForJson(df)
}
\arguments{
\item{df}{A single data frame (from a single subject/day)}
}
\value{
A list ready for jsonification
}
\description{
Return a list of the data ready for json
}
