% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxtest.R
\name{DetectionLatency}
\alias{DetectionLatency}
\title{Obtain and plot the detection latencies}
\usage{
DetectionLatency(df, return_ax = FALSE)
}
\arguments{
\item{df}{timestamped data file}
}
\value{
df the same timestamped data file
}
\description{
Obtain and plot the detection latencies
}
