% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active.R
\name{DropBoxPaths}
\alias{DropBoxPaths}
\title{Return the dropbox paths for the local and experimental computers}
\usage{
DropBoxPaths()
}
\value{
list of dropbox files (local machine and experimental machine)
}
\description{
Return the dropbox paths for the local and experimental computers
}
