% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active.R
\name{ExperimentID}
\alias{ExperimentID}
\title{Returns the experiment ID from the full experiment name}
\usage{
ExperimentID(experiment)
}
\arguments{
\item{experiment}{Full name of an experiment to identify by its ID}
}
\value{
experiment ID
}
\description{
Returns the experiment ID from the full experiment name
}
