% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active.R
\name{ExperimentName}
\alias{ExperimentName}
\title{Returns the experiment name from the experiment ID}
\usage{
ExperimentName(experiment)
}
\arguments{
\item{experimentID}{The ID for an active experiment.}
}
\value{
experiment name
}
\description{
Returns the experiment name from the experiment ID
}
