% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active.R
\name{ExperimentPath}
\alias{ExperimentPath}
\title{Returns the experiment Path from the experiment ID}
\usage{
ExperimentPath(experiment)
}
\arguments{
\item{experimentID}{The ID for an active experiment.}
}
\value{
experiment path
}
\description{
Returns the experiment Path from the experiment ID
}
