% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{NextArchive}
\alias{NextArchive}
\title{Return the next archive number}
\usage{
NextArchive(mpcpath)
}
\arguments{
\item{mpcpath}{The path to the data/mpc directory}
}
\description{
Return the next archive number
}
