% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxtest.R
\name{Pellet_Off_Before_On}
\alias{Pellet_Off_Before_On}
\title{Find where a pellet offset comes before an onset}
\usage{
Pellet_Off_Before_On(df)
}
\arguments{
\item{df}{timestamped data file}
}
\value{
df the same timestamped data file
}
\description{
Find where a pellet offset comes before an onset
}
