% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active.R
\name{ReadActiveExperimentFile}
\alias{ReadActiveExperimentFile}
\title{Read all the .999 files from an active experiment}
\usage{
ReadActiveExperimentFile(experiment)
}
\value{
The ts data frame
}
\description{
Read all the .999 files from an active experiment
}
