% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active.R
\name{ReadActiveExperimentFiles}
\alias{ReadActiveExperimentFiles}
\title{Read all the .999 files from all active experiments}
\usage{
ReadActiveExperimentFiles()
}
\value{
The ts data frame
}
\description{
Read all the .999 files from all active experiments
}
