% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active.R
\name{ReadConditionsFile}
\alias{ReadConditionsFile}
\title{Return the conditions file}
\usage{
ReadConditionsFile(File)
}
\arguments{
\item{File}{The conditions file to load}
}
\value{
The loaded conditions file
}
\description{
Return the conditions file
}
