% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{ReadExperimentJson}
\alias{ReadExperimentJson}
\title{Read in an entire experiment worth of json data}
\usage{
ReadExperimentJson(experiment)
}
\arguments{
\item{experiment}{An experiment ID}
}
\description{
Read in an entire experiment worth of json data
}
