% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active.R
\name{RecentExperimentActivity}
\alias{RecentExperimentActivity}
\title{Returns the number of .999 files}
\usage{
RecentExperimentActivity()
}
\value{
Count of .999 files
}
\description{
Returns the number of .999 files
}
