% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{WriteJsonDataFile}
\alias{WriteJsonDataFile}
\title{Write a single day's worth of Json Data}
\usage{
WriteJsonDataFile(df, path, eventcodes, protocol, conditions)
}
\arguments{
\item{df}{A processed data file (subject, day, json_lists)}

\item{path}{The save directory (data/json)}

\item{eventcodes}{The active event codes}

\item{protocol}{The active protocol}

\item{conditions}{The active conditions}
}
\description{
Write a single day's worth of Json Data
}
