% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{WriteJsonDataFiles}
\alias{WriteJsonDataFiles}
\title{Write an experiment's Json files, and archive the mpc files}
\usage{
WriteJsonDataFiles(experiment)
}
\arguments{
\item{experiment}{An active experiment ID}
}
\description{
Write an experiment's Json files, and archive the mpc files
}
