% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{WriteJsonDataFilesFromInactiveExperiment}
\alias{WriteJsonDataFilesFromInactiveExperiment}
\title{Write an experiment's Json files, after the experiment went inactive.}
\usage{
WriteJsonDataFilesFromInactiveExperiment(path)
}
\arguments{
\item{path}{A path to a valid experiment folder (most likely inactive.)}
}
\description{
Write an experiment's Json files, after the experiment went inactive.
}
