% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macros.R
\name{WriteMacros}
\alias{WriteMacros}
\title{Write the macros for a conditions file}
\usage{
WriteMacros(conditions)
}
\arguments{
\item{conditions}{A conditions file}
}
\value{
None
}
\description{
Write the macros for a conditions file
}
