% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macros.R
\name{WriteSubjectMacros}
\alias{WriteSubjectMacros}
\title{Write a single subject macro}
\usage{
WriteSubjectMacros(df)
}
\arguments{
\item{df}{Single row in a conditions file}
}
\value{
TRUE
}
\description{
Write a single subject macro
}
